% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allreefs-data.R
\docType{data}
\name{allreefs}
\alias{allreefs}
\title{all reefs data}
\format{
"sf" "data.frame"
}
\source{
https://github.com/WCS-Marine/local-reef-pressures
}
\usage{
data(allreefs)
}
\description{
Data from Andrello et al., 2022
}
\references{
Andrello, M., Darling, E.S., Wenger, A., Suárez‐Castro, A.F., Gelfand, S. and Ahmadia, G.N., 2022. A global map of human pressures on tropical coral reefs. Conservation Letters, 15(1), p.e12858.
}
\keyword{datasets}
