% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etopo-data.R
\docType{data}
\name{etopo}
\alias{etopo}
\title{ETOPO data}
\format{
RasterLayer
}
\source{
https://www.ncei.noaa.gov/products/etopo-global-relief-model
}
\usage{
data(etopo)
}
\description{
Data from ETOTPO
}
\references{
...
}
\keyword{datasets}
