% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBiogeographicVariablesForARMS().R
\name{GetBiogeographicVariablesForARMS}
\alias{GetBiogeographicVariablesForARMS}
\title{Get biographic variables}
\usage{
GetBiogeographicVariablesForARMS(
  dat,
  shallowOceanBuffer = 50000,
  shallowDepth = 200
)
}
\arguments{
\item{dat}{inout dataframe}

\item{shallowOceanBuffer}{the radius of the circle within which to calculate the proportion of shallow ocean (metres). default 50km}

\item{shallowDepth}{the depth of ocean up to which it is considered shallow, default 200m}
}
\value{
A dataframe containing the ARMs identifier, area, and distance. Also contains the LatLong of the ARMS and the id of the continental shelf which was found to be closest
}
\description{
This function accepts a dataframe with columns of 'ARMS' 'Site_Longitude' and 'Site_Latitude'
and returns a dataframe with the seamount area and isolation associated with each ARMS unit.
}
