% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData-data.R
\docType{data}
\name{DemoData}
\alias{DemoData}
\title{DemoData data}
\format{
dataframe
}
\source{
NA
}
\usage{
data(DemoData)
}
\description{
Demonstration data to test package
}
\examples{
GetBiogeographicVariablesForARMS(DemoData)

}
\references{
...
}
\keyword{datasets}
